% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bic_function.R
\name{CalculateBIC}
\alias{CalculateBIC}
\title{Calculate BIC values and generate the corresponding clustering results}
\usage{
CalculateBIC(
  times,
  oridata,
  rangeval = c(0, 1),
  nknots = 30,
  order = 4,
  nu = 2,
  tau = 3,
  K0 = 6,
  rep_num = 100,
  kappa = 1,
  eps_outer = 1e-04,
  max_iter = 100,
  lambda1,
  lambda2_seq,
  lambda3_seq
)
}
\arguments{
\item{times}{A list with n elements, each of which contains a sequence of time points.}

\item{oridata}{A list with n elements, each of wich contains a sequence of observations corresponding to the times.}

\item{rangeval}{A numeric vector of length 2 defining the interval over which the functional data object can be evaluated; default value is c(0,1).}

\item{nknots}{An integer specifying the number of equally spaced knots; default value is 30.}

\item{order}{An integer specifying the order of B-splines, which is one higher than their degree. The default of 4 gives cubic splines.}

\item{nu}{A non-negative integer defining an order of a derivative. The default of 2 corresponds to measuring the roughness of a function by its integrated curvature.}

\item{tau}{A positive double parameter that controls the concavity of mcp function. The default value is 3.}

\item{K0}{The pre-specified number of clusters; default value is 6.}

\item{rep_num}{The number of replicates for K-means; default value is 100.}

\item{kappa}{A positive penalty parameter for ADMM algorithm. The default value is 1.}

\item{eps_outer}{The convergence tolerance and its default value is 0.0001.}

\item{max_iter}{The maximum number of iterations.}

\item{lambda1}{The tuning parameter for roughness of estimated functions.}

\item{lambda2_seq}{A sequence of tuning parameters for sparsity, that is, used to bring the estimated individual functions closer to the estimated center functions.}

\item{lambda3_seq}{A sequence of tuning parameters for clustering, that is, used to encourage similar estimated center functions to be the same, automatically identifying the number of clusters.}
}
\value{
A matrix, each element of which is a list of the following items.
\itemize{
\item{\code{funlc_result} A list of local clustering results.}
\item{\code{bic_val} The corresponding bic value.}
}
}
\description{
\code{CalculateBIC} generates a matrix list of local clustering estimates and the corresponding BIC values based on the candidate lambda2 and lambda3 sequences.

Dimensions of various arguments are as follows:
\itemize{
\item{n is the number of individuals}
\item{p = \code{nknots} + \code{order}, the number of basis functions}
}
}
\seealso{
\code{\link{LFC.fit}} for local clustering.
}
