% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjacency_function.R
\name{CreateAdjacency}
\alias{CreateAdjacency}
\title{Create adjacency matrix from the pairwise difference matrix}
\usage{
CreateAdjacency(V, n)
}
\arguments{
\item{V}{Matrix of centroid differences}

\item{n}{Number of points to cluster}
}
\description{
\code{CreateAdjacency} creates an n-by-n sparse adjacency matrix from the matrix of centroid differences.
}
