% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialization_function.R
\name{GenerateInitialCoefficients}
\alias{GenerateInitialCoefficients}
\title{Generate the initial smoothed B-spline coefficients and the centered coefficients}
\usage{
GenerateInitialCoefficients(
  times,
  oridata,
  rangeval = c(0, 1),
  nknots = 30,
  order = 4,
  nu = 2,
  lambda1 = 1e-05,
  K0 = 6,
  rep_num = 100
)
}
\arguments{
\item{times}{A list with n elements, each of which contains a sequence of time points.}

\item{oridata}{A list with n elements, each of wich contains a sequence of observations corresponding to the times.}

\item{rangeval}{A numeric vector of length 2 defining the interval over which the functional data object can be evaluated; default value is c(0,1).}

\item{nknots}{An integer specifying the number of equally spaced knots; default value is 30.}

\item{order}{An integer specifying the order of B-splines, which is one higher than their degree. The default of 4 gives cubic splines.}

\item{nu}{A non-negative integer defining an order of a derivative. The default of 2 corresponds to measuring the roughness of a function by its integrated curvature.}

\item{lambda1}{The tuning parameter for roughness of estimated functions.}

\item{K0}{The pre-specified number of clusters; default value is 6.}

\item{rep_num}{The number of replicates for K-means; default value is 100.}
}
\value{
\code{Beta_ini} The p * n matrix, whose columns correspond to individuals.

\code{Alpha_ini} The p * \code{K0} matrix, whose columns correspond to centers.
}
\description{
\code{GenerateInitialCoefficients} generates the initial B-spline coefficients.

Dimensions of various arguments are as follows:
\itemize{
\item{n is the number of individuals}
\item{p = \code{nknots} + \code{order}, the number of basis functions}
}
}
