% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{Plot.LFC}
\alias{Plot.LFC}
\alias{plot.individual.localfclust}
\alias{plot.center.localfclust}
\title{Plot the results of  the LFC method}
\usage{
\method{plot}{individual.localfclust}(model.lfc)

\method{plot}{center.localfclust}(model.lfc)
}
\arguments{
\item{model.lfc}{The output of \code{\link{LFC.fit}}.}
}
\value{
The ggplot2 object.
}
\description{
Provides plots of the classified curves and estimated cluster mean functions, respectively.
}
