% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proximal_functions.R
\name{proximal}
\alias{proximal}
\alias{ProximalMCP}
\alias{ProximalAverageMCP}
\title{Proximal mapping functions}
\usage{
ProximalMCP(x, varsigma, tau = 3, lambda)

ProximalAverageMCP(x, l, r, varsigma, tau = 3, lambda)
}
\arguments{
\item{x}{The input vector.}

\item{varsigma}{A positive double scaling parameter.}

\item{tau}{A positive double parameter that controls the concavity of mcp function. The default value is 3.}

\item{lambda}{The tuning parameter for mcp function.}

\item{l}{An integer taking values from \{1,...,L\} and L is the number of subintervals.}

\item{r}{An integer specifying the degree of B-splines.}
}
\value{
A vector with the same dimensions as \code{x}.
}
\description{
The proximal (average) mapping of mcp function.
}
