% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_simulation.R
\name{SimulateData}
\alias{SimulateData}
\alias{SimulateExample1}
\alias{SimulateExample2}
\alias{SimulateExample3}
\alias{SimulateExample4}
\title{Simulate data for local functional clustering}
\usage{
SimulateExample1(seed, setpoint1 = 0.6, n = 100, n0 = 100, sigma = 0.75)

SimulateExample2(seed, n = 100, n0 = 100, sigma = 0.5)

SimulateExample3(seed, n = 100, n0 = 100, sigma = 0.5)

SimulateExample4(seed, n = 100, n0 = 100, sigma = 0.75)
}
\arguments{
\item{seed}{An integer specifying the random seed used for random data generation.}

\item{setpoint1}{A positive double parameter that sets the location of the startpoint for heterogeneous subinterval in Example 1.}

\item{n}{An integer specifying the number of functions.}

\item{n0}{An integer specifying the number of repeated measurements.}

\item{sigma}{A positive double parameter that controls the variance of random errors}
}
\value{
A list of the following items.
\itemize{
\item \code{location}: A list of n vector and each vector collects the observed time points of the sequence data.
\item \code{data.lst}: A list of n vectors and each vector collects the observed replications.
\item \code{clus.true}: True cluster membership vector.
}
}
\description{
Generate synthetic data for Examples 1-4 as in the simulation study (Chen et al., 2024).
}
